Refactor model file names and update CORS configuration, Marker se propíše do databáze ale je problém dostat ho z databáze

+++ proj_tt/api/Models/GPSPoint.cs
@@ -8,8 +8,8 @@ namespace TT_API.Models {
         [Key]
         public int PointID { get; set; }
         public int IdMap { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
 
     }
 }
